requirements.txt
grpcio==1.53.0 
grpcio-tools==1.53.0 
mysql-connector-python==8.0.33 
